+++ features/suite_test.go
@@ -339,6 +339,7 @@ func initializeScenario(sc *godog.ScenarioContext) {
 	initializeInfoSteps(sc)
 	initializeEventsSteps(sc)
 	initializeTurnsSteps(sc)
+	initializeTurnsViewSteps(sc)
 	initializeAttachSteps(sc)
 	initializeContextSteps(sc)
 	initializeSandboxEnvSteps(sc)